MultiSite: Remove Website functionality Additional task, to check db user exist or not.
git-svn-id: http://svn.contrexxlabs.com/contrexx/trunk@32813 5ce06e5d-2b28-0410-aa6a-b3cd9cf52d8a

+++ core_modules/MultiSite/Controller/XamppController.class.php
@@ -101,10 +101,12 @@ class XamppController implements \Cx\Core_Modules\MultiSite\Controller\DbControl
      * @throws MultiSiteDbException On error
      */
     public function removeDbUser(\Cx\Core\Model\Model\Entity\DbUser $dbUser){
-        $isUserDeleted = $this->db->execute("DROP USER '".$dbUser->getName()."'@'localhost'");
-        if (!$isUserDeleted) {
-            throw new \Exception("Query failed: \ DROP USER '".$dbUser->getName()."'@'localhost'" . $this->db->ErrorMsg());
-        }    
+        if ($dbUser->getName()) {
+            $isUserDeleted = $this->db->execute("DROP USER '".$dbUser->getName()."'@'localhost'");
+            if (!$isUserDeleted) {
+                throw new \Exception("Query failed: \ DROP USER '".$dbUser->getName()."'@'localhost'" . $this->db->ErrorMsg());
+            }    
+        } 
     }
     
     /**
@@ -113,10 +115,12 @@ class XamppController implements \Cx\Core_Modules\MultiSite\Controller\DbControl
      * @throws MultiSiteDbException On error
      */
     public function removeDb(\Cx\Core\Model\Model\Entity\Db $db){
-        $isDbCreated = $this->db->execute("DROP DATABASE `".$db->getName()."`");
-        if (!$isDbCreated) {
-            throw new \Exception('Query failed: \'DROP DATABASE `'.$db->getName().'`\', ' . $this->db->ErrorMsg());
-        }    
+        if ($db->getName()) {
+            $isDbCreated = $this->db->execute("DROP DATABASE `".$db->getName()."`");
+            if (!$isDbCreated) {
+                throw new \Exception('Query failed: \'DROP DATABASE `'.$db->getName().'`\', ' . $this->db->ErrorMsg());
+            }  
+        }
     }
     
      /**
